+++ pom.xml
@@ -61,12 +61,6 @@
 	        <version>1.5.5.Final</version>
 	        <scope>provided</scope>
     	</dependency>
-    	<dependency>
-	        <groupId>org.jacoco</groupId>
-	        <artifactId>jacoco-maven-plugin</artifactId>
-	        <version>0.8.11</version>
-	        <scope>test</scope>
-	    </dependency>
 	</dependencies>
 
 	<build>
@@ -83,26 +77,6 @@
 					</excludes>
 				</configuration>
 			</plugin>
-			<plugin>
-	            <groupId>org.jacoco</groupId>
-	            <artifactId>jacoco-maven-plugin</artifactId>
-	            <version>0.8.12</version>
-	            <executions>
-	                <execution>
-	                    <id>prepare-agent</id>
-	                    <goals>
-	                        <goal>prepare-agent</goal>
-	                    </goals>
-	                </execution>
-	                <execution>
-	                    <id>report</id>
-	                    <phase>prepare-package</phase>
-	                    <goals>
-	                        <goal>report</goal>
-	                    </goals>
-	                </execution>
-	            </executions>
-	        </plugin>
 		</plugins>
 	</build>
