+++ src/main/java/com/comercio/electronico/api/products/application/core/domain/exception/NoPricesFoundException.java
@@ -3,7 +3,6 @@ package com.comercio.electronico.api.products.application.core.domain.exception;
 /**
  * Excepción lanzada cuando no se encuentran los precios consultados.
  */
-@SuppressWarnings("serial")
 public class NoPricesFoundException extends RuntimeException {
 
     /**
